# Faster trie node encoding updated